Enum libsystemd_sys::c_void

#[repr(u8)]

pub enum c_void {
    // some variants omitted
}

Type used to construct void pointers for use with C.

This type is only useful as a pointer target. Do not use it as a return type for FFI functions which have the void return type in C. Use the unit type () or omit the return type instead.