Crate libsystemd_sys

Contains definitions for low-level bindings.

Most of this module is Rust versions of the systemd headers. The goal of this crate is to make it unattractive to ever use the FFI directly, but it's there if you need it.

Items in this module corresponding to systemd functions are well-documented by the systemd man pages.

Modules

bus

daemon

event

id128

journal

login

Structs

const_iovec

iovec

siginfo_t

signalfd_siginfo

Enums

c_void
Equivalent to C's void type when used as a pointer.

Functions

array_to_iovecs

Type Definitions

c_char
Equivalent to C's char type.

c_int
Equivalent to C's signed int (int) type.

c_uint
Equivalent to C's unsigned int type.

clockid_t

gid_t

pid_t

size_t

uid_t