Crate libsystemd_sys

Contains definitions for low-level bindings.

Most of this module is Rust versions of the systemd headers. The goal of this crate is to make it unattractive to ever use the FFI directly, but it's there if you need it.

Items in this module corresponding to systemd functions are well-documented by the systemd man pages.

Reexports

pub use daemon::*;

Modules

bus

daemon

event

id128

Structs

const_iovec

iovec

Enums

c_void
Type used to construct void pointers for use with C.

Constants

SD_JOURNAL_CURRENT_USER

SD_JOURNAL_LOCAL_ONLY

SD_JOURNAL_RUNTIME_ONLY

SD_JOURNAL_SYSTEM

Functions

array_to_iovecs

sd_journal_add_conjunction

sd_journal_add_disjunction

sd_journal_add_match

sd_journal_close

sd_journal_enumerate_data

sd_journal_enumerate_unique

sd_journal_flush_matches

sd_journal_get_catalog

sd_journal_get_catalog_for_message_id

sd_journal_get_cursor

sd_journal_get_cutoff_monotonic_usec

sd_journal_get_cutoff_realtime_usec

sd_journal_get_data

sd_journal_get_data_threshold

sd_journal_get_events

sd_journal_get_fd

sd_journal_get_monotonic_usec

sd_journal_get_realtime_usec

sd_journal_get_timeout

sd_journal_get_usage

sd_journal_next

sd_journal_next_skip

sd_journal_open

sd_journal_previous

sd_journal_previous_skip

sd_journal_process

sd_journal_query_unique

sd_journal_reliable_fd

sd_journal_restart_data

sd_journal_restart_unique

sd_journal_seek_cursor

sd_journal_seek_head

sd_journal_seek_monotonic_usec

sd_journal_seek_realtime_usec

sd_journal_seek_tail

sd_journal_sendv

sd_journal_set_data_threshold

sd_journal_test_cursor

sd_journal_wait

Type Definitions

c_char

c_int

c_uint

gid_t

pid_t

sd_journal

size_t

uid_t