Crate libsystemd_sys

Contains definitions for low-level bindings.

Most of this module is Rust versions of the systemd headers. The goal of this crate is to make it unattractive to ever use the FFI directly, but it's there if you need it.

Items in this module corresponding to systemd functions are well-documented by the systemd man pages.

Modules

daemon

event

id128

journal

login

Structs

const_iovec

iovec

siginfo_t

signalfd_siginfo

Enums

c_void
Type used to construct void pointers for use with C.

Functions

array_to_iovecs

Type Definitions

c_char

c_int

c_uint

clockid_t

gid_t

int64_t

pid_t

size_t

uid_t

uint32_t

uint64_t